+++ cli/cli/Commands/Services/ServicesDeployCommand.cs
@@ -1,6 +1,9 @@
 ﻿using Beamable.Common;
+using Beamable.Common.Api;
+using Beamable.Common.Api.Realms;
 using Beamable.Common.BeamCli;
 using cli.Services;
+using cli.Utils;
 using Newtonsoft.Json;
 using Serilog;
 using Spectre.Console;
@@ -29,11 +32,14 @@ public class ServicesDeployCommand : AppCommand<ServicesDeployCommandArgs>,
 	private IAppContext _ctx;
 	private BeamoLocalSystem _localBeamo;
 	private BeamoService _remoteBeamo;
+	private IAliasService _aliasService;
+	private InitCommand _initCommand;
 
-	public ServicesDeployCommand() :
+	public ServicesDeployCommand(InitCommand initCommand) :
 		base("deploy",
 			"Deploys services remotely to the current realm")
 	{
+		_initCommand = initCommand;
 	}
 
 	public override void Configure()
@@ -64,11 +70,34 @@ public class ServicesDeployCommand : AppCommand<ServicesDeployCommandArgs>,
 		_ctx = args.AppContext;
 		_localBeamo = args.BeamoLocalSystem;
 		_remoteBeamo = args.BeamoService;
+		_aliasService = args.AliasService;
 
 		var isDockerRunning = await _localBeamo.CheckIsRunning();
 		if (!isDockerRunning)
 		{
 			throw new CliException("Docker is not running in this machine. Please start Docker before running this command.", Beamable.Common.Constants.Features.Services.CMD_RESULT_CODE_DOCKER_NOT_RUNNING, true);
+		}
+
+		var cid = _ctx.Cid;
+		if (!AliasHelper.IsCid(cid))
+		{
+			try
+			{
+				var aliasResolve = await _aliasService.Resolve(cid).ShowLoading("Resolving alias...");
+				cid = aliasResolve.Cid.GetOrElse(() => throw new CliException("Invalid alias"));
+				_ctx.Set(cid, _ctx.Pid, _ctx.Host);
+			}
+			catch (RequesterException)
+			{
+				AnsiConsole.WriteLine($"Organization not found for '{cid}', try again");
+				await _initCommand.Handle(new InitCommandArgs { Provider = args.Provider, saveToFile = true });
+				return;
+			}
+			catch (Exception e)
+			{
+				BeamableLogger.LogError(e.Message);
+				return;
+			}
 		}
 
 		try