+++ cli/cli/CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for storing locally content from multiple namespaces.
 - Filter out storage objects from `beam services enable` selection wizard.
 
+### Fixed
+- Validate cid and resolve alias to cid on microservice deploy.
+
 ### Changed
 - `run-nbomber` cli command accepts a json file as body for request instead of an argument.